design issues

+++ webapp/backend/src/server/services/baskets.service.js
@@ -71,7 +71,7 @@ class BasketsService {
 
       try {
         const orders = await queryRunner.query(
-          `select * from orders where user_id = $1 and status='basket' limit 1`,
+          `select * from orders where user_id = $1::int and status='basket' limit 1`,
           [user_id]
         );
         const basket_id = orders[0].id;
